backend/requirements.txt
fastapi[standard]
pillow
requests
html5lib
bs4
pytest
serpapi
google-search-results